back/face_api/requirements.txt
fastapi
uvicorn
opencv-python
dlib
numpy
imutils
Pillow
